Rotator class
Created a rotator class that will rotate the alphabet 4 times. It kind
of works, it will rotate the alphabet 26 times for the elements in the
array and then will do that 4 more times, for the letters that we want
to encrypt. Needless to say it creates a shit ton of arrays.